Fix unknown flag `bitcoin.rpcuser' error
When running the command with the --bitcoin flag, it throws the error:

`unknown flag 'bitcoin.rpcuser'`

It seems that with `lnd`, you have to specify the actual bitcoin daemon that you're connecting to, so either `btcd` or `bitcoind`. Since the tutorial assumes you're using `btcd`, the tutorial should use that flag instead.